+++ dat110-project1-tests/src/no/hvl/dat110/rpc/tests/TestVoidVoidStub.java
@@ -14,7 +14,7 @@ public class TestVoidVoidStub extends RPCLocalStub {
 		
 		byte[] request = RPCUtils.marshallVoid();
 		
-		byte[] reply = rpcclient.call(1,request);
+		byte[] reply = rpcclient.call((byte)1,request);
 		
 		RPCUtils.unmarshallVoid(reply);
